requirements.txt
asgiref==3.3.1
Django==3.1.4
django-admin-rangefilter==0.6.3
pytz==2020.4
sqlparse==0.4.1
